# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Authentication API",
        "description": "This is a sample Authentication api and store product CRUD",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "summary": "login Api",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register Api",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name .",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "c_password",
                        "in": "formData",
                        "description": "Re password.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/profile": {
            "post": {
                "summary": "My Profile Api",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "use Bearer . Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "summary": "List All Product ",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/product/store": {
            "post": {
                "summary": "Add new Product by user Api",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "use Bearer . Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Product title.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Product description .",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Product category ID.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Image product.",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/product/{product_id}": {
            "get": {
                "summary": "Show one Product by id ",
                "operationId": "getCustomerRates",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Target Product.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {}
}